Cleans up app.js and adds new file for tournament business logic
